requirements.txt
django>=1.10
django-allauth>=0.26.1
django-multiselectfield

